Revert back to using single space when wrapping headers
Turns out, the iCloud SMTP server was unable to properly read the from field when it got wrapped using a tab instead of a space, failing with a "Not one of your from addresses" error message. The decision to use tabs instead of spaces just because Mail did so for the Content-Type header (and interestingly NOT for any other headers) was a bad one. After some research, it seems like various SMTP servers wrap the content-type header with a boundary with spaces, tabs or no characters and all of them seem to be working.